genai-feat-private_assistant_v2-multimodal/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.8.2
pyyaml==6.0.1
python-dotenv==1.0.1
python-multipart==0.0.9

# UI
gradio==4.44.1
requests==2.32.3

# multimodal ingestion
pillow==10.4.0
pytesseract==0.3.10
pypdf==4.3.1
python-docx==1.1.2

# optional privacy
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354

# pin HF hub for gradio/oauth compatibility (avoid HfFolder import issues)
huggingface_hub==0.24.7
